Add utility function for formatting zulu time (now)
There are a couple of places this common operation is done in
objects.c, and it would benefit from being isolated in one function
rather than cut and pasted.

This commit creates a single function that formats the current time in
GMT for use as a packet timestamp, and a simple unit test that uses a
stub implementation of sec_now() (normally in util.c) which returns a
fixed timestamp representing 11 November 2025 16:18:00 UTC, and
verifies that the correct zulu timestamp is produced as expected.

It then makes use of this function in `Create_object_item_tx_string`
instead of being coded directly there.  I tested this integration by
creating objects in Xastir and watching the transmitted packet for
correct format.
